separate hooks into api and nats
