+++ frontend/interactEM/src/components/nodes/header.tsx
@@ -1,7 +1,7 @@
 import ErrorIcon from "@mui/icons-material/Error"
 import { Tooltip } from "@mui/material"
 import type React from "react"
-import { useOperatorEvents } from "../../hooks/useOperatorEvents"
+import { useOperatorEvents } from "../../hooks/nats/useOperatorEvents"
 import { type OperatorErrorEvent, OperatorEventType } from "../../types/events"
 
 interface OperatorHeaderProps {
