+++ frontend/interactEM/src/components/nodes/parameterupdater.tsx
@@ -16,8 +16,8 @@ import type React from "react"
 import { memo, useEffect, useState } from "react"
 import type { OperatorParameter } from "../../client"
 import { usePipelineContext } from "../../contexts/pipeline"
-import { useParameterUpdate } from "../../hooks/useParameterUpdate"
-import { useParameterValue } from "../../hooks/useParameterValue"
+import { useParameterUpdate } from "../../hooks/nats/useParameterUpdate"
+import { useParameterValue } from "../../hooks/nats/useParameterValue"
 import { useEditModeState } from "../../stores/edit"
 import type { OperatorNodeType } from "../../types/nodes"
 import ParameterInfoTooltip from "./parameterinfotooltip"
