+++ frontend/interactEM/src/components/pipelines/revisionlist.tsx
@@ -1,7 +1,7 @@
 import { Box, CircularProgress, Divider, List, Typography } from "@mui/material"
 import { useRef } from "react"
 import InfiniteScroll from "react-infinite-scroll-component"
-import { useInfinitePipelineRevisions } from "../../hooks/usePipelinesQuery"
+import { useInfinitePipelineRevisions } from "../../hooks/api/usePipelinesQuery"
 import { usePipelineStore } from "../../stores"
 import { RevisionListItem } from "./revisionlistitem"
