+++ frontend/interactEM/src/hooks/nats/useImage.ts
@@ -1,7 +1,7 @@
-import { AckPolicy, DeliverPolicy, ReplayPolicy } from "@nats-io/jetstream"
 import type { JsMsg } from "@nats-io/jetstream"
+import { AckPolicy, DeliverPolicy, ReplayPolicy } from "@nats-io/jetstream"
 import { useCallback, useMemo, useState } from "react"
-import { IMAGES_STREAM } from "../constants/nats"
+import { IMAGES_STREAM } from "../../constants/nats"
 import { useConsumeMessages } from "./useConsumeMessages"
 import { useConsumer } from "./useConsumer"
 import { useStream } from "./useStream"
