+++ frontend/interactEM/src/hooks/nats/useOperatorEvents.ts
@@ -1,8 +1,8 @@
-import { AckPolicy, DeliverPolicy, ReplayPolicy } from "@nats-io/jetstream"
 import type { JsMsg } from "@nats-io/jetstream"
+import { AckPolicy, DeliverPolicy, ReplayPolicy } from "@nats-io/jetstream"
 import { useCallback, useMemo, useState } from "react"
-import { OPERATORS_STREAM } from "../constants/nats"
-import type { OperatorEvent } from "../types/events"
+import { OPERATORS_STREAM } from "../../constants/nats"
+import type { OperatorEvent } from "../../types/events"
 import { useConsumeMessages } from "./useConsumeMessages"
 import { useConsumer } from "./useConsumer"
 import { useStream } from "./useStream"
