+++ frontend/interactEM/src/hooks/nats/useParameterUpdate.ts
@@ -3,9 +3,9 @@ import {
   useMutation,
   useQueryClient,
 } from "@tanstack/react-query"
-import type { OperatorParameter } from "../client"
-import { PARAMETERS_QUERYKEY, PARAMETERS_STREAM } from "../constants/nats"
-import { useNats } from "../contexts/nats"
+import type { OperatorParameter } from "../../client"
+import { PARAMETERS_QUERYKEY, PARAMETERS_STREAM } from "../../constants/nats"
+import { useNats } from "../../contexts/nats"
 
 export const useParameterUpdate = (
   operatorID: string,
