+++ frontend/interactEM/src/hooks/nats/useParameterValue.ts
@@ -1,7 +1,10 @@
-import { AckPolicy, DeliverPolicy, ReplayPolicy } from "@nats-io/jetstream"
 import type { JsMsg } from "@nats-io/jetstream"
+import { AckPolicy, DeliverPolicy, ReplayPolicy } from "@nats-io/jetstream"
 import { useCallback, useMemo, useState } from "react"
-import { PARAMETERS_STREAM, PARAMETERS_UPDATE_STREAM } from "../constants/nats"
+import {
+  PARAMETERS_STREAM,
+  PARAMETERS_UPDATE_STREAM,
+} from "../../constants/nats"
 import { useConsumeMessages } from "./useConsumeMessages"
 import { useConsumer } from "./useConsumer"
 import { useStream } from "./useStream"
