+++ frontend/interactEM/src/pages/composerpage.tsx
@@ -7,9 +7,9 @@ import ComposerPipelineFlow from "../components/composerpipelineflow"
 import { OperatorMenu } from "../components/operatormenu"
 import { PipelineHud } from "../components/pipelines/hud"
 import { usePipelineContext } from "../contexts/pipeline"
-import { useActivePipeline } from "../hooks/useActivePipeline"
-import { useAgents } from "../hooks/useAgents"
-import useOperators from "../hooks/useOperators"
+import { useActivePipeline } from "../hooks/api/useActivePipeline"
+import useOperators from "../hooks/api/useOperators"
+import { useAgents } from "../hooks/nats/useAgents"
 import { usePipelineStore } from "../stores"
 import { useEditModeState } from "../stores/edit"
